+++ app/helpers/paxful_engine/application_helper.rb
@@ -1,4 +1,10 @@
 module PaxfulEngine
   module ApplicationHelper
+
+    def highlight_status(trade)
+      status = trade.status.downcase
+      status == "successful" ? "text-success" : "text-danger"
+    end
+
   end
 end